WIP: tests non-functional
